requirements.txt
PyAudio==0.2.14
python-docx==1.1.2
reportlab==4.2.5
SpeechRecognition==3.11.0
streamlit==1.40.0

